created url security and comment CRUD security

+++ Laravel/laravelwebsite/resources/views/comments/delete.blade.php
@@ -1,8 +1,14 @@
 @extends('layouts.app')
 
 @section('content')
-    <h1>Are you sure you want to delete this comment?<h1>    
-    {{Form::open(['route' => ['comments.destroy', $comment->id], 'method' => 'DELETE'])}}       
-        {{Form::submit('Confirm delete', ['class' => 'btn btn-block, btn-danger'])}}        
-    {{Form::close()}}
+    @if(!Auth::guest()) 
+        @if(Auth::user()->id == $comment->user_id)
+        <h1>Are you sure you want to delete this comment?<h1>    
+        {{Form::open(['route' => ['comments.destroy', $comment->id], 'method' => 'DELETE'])}}       
+            {{Form::submit('Confirm delete', ['class' => 'btn btn-block, btn-danger'])}}        
+        {{Form::close()}}
+        @else
+            <p>YOU DO NOT HAVE THE AUTHORITY TO DO THIS</p>
+        @endif 
+    @endif
 @endsection
